+++ src/components/profile/Profile.tsx
@@ -8,10 +8,10 @@ import { defaultImages } from '@/assets'
 import i18n from '@/locales'
 import { Permissions } from '@/models/enums/Permissions'
 import { ProfileMenuType } from '@/models/enums/ProfileMenuType'
-import { ProfileImageState } from '@/recoils/ProfileImageState'
 import { createCancleConfirmAlert } from '@/services/Alert'
 import { checkPermission } from '@/services/PermissionService'
 import { handleCameraOpen, handleImageSelect } from '@/services/imagePicker'
+import { ProfileImageState } from '@/state/ProfileImageState'
 import { colors } from '@/styles/colors'
 
 import { Container, EmptyProfile, ProfileImage, UploadButton } from './Profile.style'
